requirements.txt
torch>=2.0.0
torchvision>=0.15.0
mlflow>=2.10.0
pyyaml>=6.0
pillow>=10.0.0
pycocotools>=2.0.7
